separation of concerns for dashboard component and mulitple colors for chart
