+++ frontend/src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
-  // apiUrl: 'http://localhost:5000/api' // Ensured correct API base URL
-  apiUrl: 'https://spiltwise-backend.onrender.com/api' // Ensured correct API base URL
+  apiUrl: 'http://localhost:5000/api' // Ensured correct API base URL
+  // apiUrl: 'https://spiltwise-backend.onrender.com/api' // Ensured correct API base URL
 };
